+++ src/app/index.app.js
@@ -24,16 +24,6 @@
         'LocalStorageModule',
         'angularScreenfull',
         'ui.ace',
-
-        // 'ui.grid',
-        // 'ui.grid.autoResize',
-        // 'ui.grid.resizeColumns',
-        // 'ui.grid.moveColumns',
-        // 'ui.grid.exporter',
-        // 'ui.grid.selection',
-        // 'ngAnimate',
-        // 'gridster',
-
         'ngScrollbars',
 
         'angularResizable',